all amount changed to decimal (".")

+++ app/src/main/java/com/ekspensify/app/data/model/requestModel/AccountRequestModel.kt
@@ -15,5 +15,5 @@ data class AccountRequestModel(
 
     @SerializedName("balance")
     @Expose
-    val balance: Int? = null,
+    val balance: Double? = null,
 ) : Serializable
