+++ app/src/main/java/com/ekspensify/app/data/model/requestModel/TransactionRequestModel.kt
@@ -19,7 +19,7 @@ data class TransactionRequestModel(
 
     @SerializedName("amount")
     @Expose
-    var amount: Int? = null,
+    var amount: Double? = null,
 
     @SerializedName("note")
     @Expose
